[#2524] Separate BarrowmanCalculator into DragCalculator and StabilityCalculator
